+++ lib/manifest/sass.rb
@@ -13,13 +13,27 @@ module Statics
     end
 
     def build
-      engine = ::SassC::Engine.new(File.read(path), syntax: :sass, style: :compressed)
+      if exists?
+        engine = ::SassC::Engine.new(File.read(path), syntax: :sass, style: :compressed)
 
-      File.write "build/#{name}.css", engine.render
+        File.write "build/#{name}.css", engine.render
+      end
+    end
+
+    def exists?
+      File.exist? path
     end
 
     def print(indentation)
-      puts "#{" " * indentation}↳Sass: #{path}"
+      puts "#{" " * (indentation * 3)}└─ Sass: #{path} #{valid}"
+    end
+
+    def valid
+      if exists?
+        Colors.in_green("✔")
+      else
+        Colors.in_red("✘")
+      end
     end
   end
 end
